+++ release-notes.md
@@ -1,5 +1,16 @@
 # Mercury 13.15.0 #
 
+## Update eSearch permissions ##
+
+- Issue Type: Task
+- Issue ID: PLAT-8618
+
+### Configuration ###
+None.
+
+### Deployment scripts ###
+      php /opt/kaltura/app/deployment/updates/scripts/add_permissions/2018_02_28_update_esearch_permissions.php
+
 ## Add new Drop Folder Type ##
 
 - Issue Type: New Feature